refactor(nvim): statusline refactors!
