+++ config/nvim/lua/mega/plugins/paq.lua
